clean up code and refactor everything better look
