+++ src/config/passport.js
@@ -1,20 +1,13 @@
 /* eslint no-param-reassign: ["error", { "props": false }]*/
 // disabling eslint rule since no "reassign" happening ->  creating properties!
-
 import { Strategy } from 'passport-twitter';
 
+import constants from './constants';
 import User from '../models/user';
 
 export default function (passport) {
   passport.use(
-		new Strategy(
-  {
-    consumerKey: process.env.TWITTER_CONSUMER_KEY,
-    consumerSecret: process.env.TWITTER_CONSUMER_SECRET,
-    callbackURL: 'http://127.0.0.1:3000/auth/twitter/callback',
-    passReqToCallback: true,
-  },
-			(req, token, tokenSecret, profile, cb) => {
+		new Strategy(constants.TWITTER_STRATEGY, (req, token, tokenSecret, profile, cb) => {
   process.nextTick(() => {
     if (!req.user) {
       User.findOne({ 'twitter.id': profile.id }, (err, user) => {
@@ -32,7 +25,7 @@ export default function (passport) {
           return cb(null, user);
         }
 
-							// if no user is found create one
+						// if no user is found create one
         const newUser = new User();
 
         newUser.twitter.id = profile.id;
@@ -46,7 +39,7 @@ export default function (passport) {
         });
       });
     } else {
-						// when user already exists and is logged in
+					// when user already exists and is logged in
       const user = req.user;
 
       user.twitter.id = profile.id;
@@ -60,8 +53,7 @@ export default function (passport) {
       });
     }
   });
-},
-		),
+}),
 	);
 
   passport.serializeUser((user, cb) => {
