+++ src/models/poll.js
